# docs/output/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: 예시 문서입니다.
  description: 예시 API
  title: Example API
  version: 1.0.0
servers:
- description: 개발용 로컬 서버
  url: http://localhost:8080
- description: 개발용 목 서버
  url: http://localhost:8081
...
paths:
...
  /api/v1/api-specs/by-project/{scrudProjectId}:
    get:
      description: 중간 매핑을 통해 ScrudProject ID로 연결된 모든 API 스펙 버전을 조회합니다.
      operationId: getApiSpecsByScrudProjectId
      parameters:
      - description: Scrud 프로젝트 ID
        explode: false
        in: path
        name: scrudProjectId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSpecListResponse'
          description: API 스펙 목록 조회 성공
      summary: Scrud 프로젝트 ID로 API 스펙 버전 목록 조회
      tags:
      - ApiSpec
  /api/v1/api-specs/{apiSpecVersionId}:
    delete:
      description: 단일 API 스펙과 관련 필드를 삭제합니다.
      operationId: deleteApiSpec
      parameters:
      - explode: false
        in: path
        name: apiSpecVersionId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: 성공적으로 처리되었습니다
      summary: API 스펙 삭제
      tags:
      - ApiSpec
    get:
      description: 단일 API 스펙과 관련 필드를 조회합니다.
      operationId: getApiSpecById
      parameters:
      - explode: false
        in: path
        name: apiSpecVersionId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSpecVersionResponse'
          description: 단일 API 스펙 조회 성공
      summary: API 스펙 상세 조회
      tags:
      - ApiSpec
    put:
      description: 단일 API 스펙과 필드를 전체 수정합니다.
      operationId: updateApiSpec
      parameters:
      - explode: false
        in: path
        name: apiSpecVersionId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/ApiSpecUpdateRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSpecVersionResponse'
          description: API 스펙 수정 성공
      summary: API 스펙 수정
      tags:
      - ApiSpec
  /api/v1/api-specs:
    post:
      description: 새로운 API 스펙과 필드를 등록합니다.
      operationId: createApiSpec
      requestBody:
        $ref: '#/components/requestBodies/ApiSpecCreateRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiSpecVersionCreatedResponse'
          description: API 스펙 생성 성공
      summary: API 스펙 생성
      tags:
      - ApiSpec
  /api/v1/api-specs/api/{apiSpecId}:
    patch:
      description: API status 상태를 수정합니다.
      operationId: updateApiSpecStatus
      parameters:
      - explode: false
        in: path
        name: apiSpecId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/ApiSpecStatusRequest'
      responses:
        "200":
          description: 성공적으로 처리되었습니다
      summary: API 스펙 status 수정
      tags:
      - ApiSpec
...
  /api/v1/posts:
    get:
      operationId: getPostList
      parameters:
      - description: "요청할 페이지 번호 (기본값: 0)"
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      - description: "페이지당 항목 수 (기본값: 10)"
        in: query
        name: size
        required: false
        schema:
          default: 10
          maximum: 100
          minimum: 1
          type: integer
      - description: 정렬 기준 필드
        in: query
        name: sort
        required: false
        schema:
          $ref: '#/components/schemas/PostSortEnumDto'
      - description: 정렬 방향
        in: query
        name: order
        required: false
        schema:
          $ref: '#/components/schemas/PostOrderEnumDto'
      - description: 검색 키워드
        in: query
        name: keyword
        required: false
        schema:
          type: string
      - description: 검색 대상 필드
        in: query
        name: type
        required: false
        schema:
          $ref: '#/components/schemas/SearchTypeEnumDto'
      - description: 카테고리 ID
        in: query
        name: categoryId
        required: false
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPostListResponse'
          description: 게시글 목록 조회 응답
      summary: 게시글 목록 조회
      tags:
      - Post
    post:
      operationId: createPost
      requestBody:
        $ref: '#/components/requestBodies/CreatePostRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatePostResponse'
          description: 게시글 생성 성공 응답
      summary: 게시글 작성
      tags:
      - Post
  /api/v1/posts/{postId}:
    delete:
      operationId: deletePost
      parameters:
      - in: path
        name: postId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: 성공적으로 처리되었습니다
      summary: 게시글 삭제
      tags:
      - Post
    get:
      operationId: getPostById
      parameters:
      - in: path
        name: postId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostDetailResponse'
          description: 게시글 상세 조회 응답
      summary: 게시글 상세 조회
      tags:
      - Post
    put:
      operationId: updatePost
      parameters:
      - in: path
        name: postId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        $ref: '#/components/requestBodies/UpdatePostRequest'
      responses:
        "200":
          description: 성공적으로 처리되었습니다
      summary: 게시글 수정
      tags:
      - Post
  /api/v1/posts/{postId}/vote:
    post:
      operationId: votePost
      parameters:
      - in: path
        name: postId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        $ref: '#/components/requestBodies/PostVoteRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoteResponse'
          description: 게시글 추천/비추천 응답
      summary: 게시글 추천/비추천
      tags:
      - Post
  /api/v1/posts/{postId}/status:
    patch:
      operationId: updatePostStatus
      parameters:
      - in: path
        name: postId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        $ref: '#/components/requestBodies/UpdatePostStatusRequest'
      responses:
        "200":
          description: 성공적으로 처리되었습니다
      summary: 게시글 상태 변경
      tags:
      - Post
...
components:
...
  requestBodies:
...
    ApiSpecUpdateRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSpecVersionUpdateRequest'
      description: 단일 API 스펙과 필드를 전체 생성합니다.
      required: true
    ApiSpecCreateRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSpecVersionCreateRequest'
      description: 단일 API 스펙과 필드를 전체 생성합니다.
      required: true
    ApiSpecStatusRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiSpecVersionStatusRequest'
      description: API 스펙 상태 변경
      required: true
    CreatePostRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CreatePostRequest'
      description: 게시글을 작성합니다.
      required: true
    UpdatePostRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdatePostRequest'
      description: 게시글을 수정합니다.
      required: true
    PostVoteRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/PostVoteRequest'
      description: 게시글 추천/비추천 요청입니다.
      required: true
    UpdatePostStatusRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpdatePostStatusRequest'
      description: 게시글 상태를 변경합니다.
      required: true
...
  schemas:
...
    ApiSpecListResponse:
      example:
        content:
        - summary: 예시 스펙입니다.
          description: 이 API는 예시 데이터를 처리합니다.
          apiSpecId: 10
          httpMethod: GET
          version: 1
          pathParameters: "[{\"name\": \"exampleId\", \"type\": \"string\", \"required\"\
            : true}]"
          endpoint: "/api/v1/examples/{id}"
          apiGroup: example
          queryParameters: "[{\"name\": \"id\", \"type\": \"integer\", \"required\"\
            : true}]"
          requestBody: "{\"name\":\"홍길동\"}"
          response: "{\"message\":\"성공\"}"
          apiSpecStatus: AI_GENERATED
          apiSpecVersionId: 101
        - summary: 예시 스펙입니다.
          description: 이 API는 예시 데이터를 처리합니다.
          apiSpecId: 10
          httpMethod: GET
          version: 1
          pathParameters: "[{\"name\": \"exampleId\", \"type\": \"string\", \"required\"\
            : true}]"
          endpoint: "/api/v1/examples/{id}"
          apiGroup: example
          queryParameters: "[{\"name\": \"id\", \"type\": \"integer\", \"required\"\
            : true}]"
          requestBody: "{\"name\":\"홍길동\"}"
          response: "{\"message\":\"성공\"}"
          apiSpecStatus: AI_GENERATED
          apiSpecVersionId: 101
      properties:
        content:
          items:
            $ref: '#/components/schemas/ApiSpecResponse'
          type: array
      type: object
    ApiSpecResponse:
      example:
        summary: 예시 스펙입니다.
        description: 이 API는 예시 데이터를 처리합니다.
        apiSpecId: 10
        httpMethod: GET
        version: 1
        pathParameters: "[{\"name\": \"exampleId\", \"type\": \"string\", \"required\"\
          : true}]"
        endpoint: "/api/v1/examples/{id}"
        apiGroup: example
        queryParameters: "[{\"name\": \"id\", \"type\": \"integer\", \"required\"\
          : true}]"
        requestBody: "{\"name\":\"홍길동\"}"
        response: "{\"message\":\"성공\"}"
        apiSpecStatus: AI_GENERATED
        apiSpecVersionId: 101
      properties:
        apiSpecId:
          example: 10
          format: int64
          type: integer
        apiSpecVersionId:
          example: 101
          format: int64
          type: integer
        endpoint:
          example: "/api/v1/examples/{id}"
          type: string
        apiGroup:
          example: example
          type: string
        version:
          example: 1
          type: integer
        summary:
          example: 예시 스펙입니다.
          type: string
        description:
          example: 이 API는 예시 데이터를 처리합니다.
          type: string
        apiSpecStatus:
          enum:
          - AI_GENERATED
          - AI_VISUALIZED
          - USER_COMPLETED
          example: AI_GENERATED
          type: string
        httpMethod:
          description: "HTTP 메서드 (GET, POST, PUT, PATCH, DELETE)"
          enum:
          - GET
          - POST
          - PUT
          - PATCH
          - DELETE
          type: string
        requestBody:
          example: "{\"name\":\"홍길동\"}"
          type: string
        queryParameters:
          example: "[{\"name\": \"id\", \"type\": \"integer\", \"required\": true}]"
          type: string
        pathParameters:
          example: "[{\"name\": \"exampleId\", \"type\": \"string\", \"required\"\
            : true}]"
          type: string
        response:
          example: "{\"message\":\"성공\"}"
          type: string
      type: object
    ApiSpecVersionResponse:
      example:
        summary: 예시 스펙입니다.
        pathParameters: "[{\"name\": \"exampleId\", \"type\": \"string\", \"required\"\
          : true}]"
        endpoint: "/api/v1/examples/{id}"
        apiGroup: example
        queryParameters: "[{\"name\": \"id\", \"type\": \"integer\", \"required\"\
          : true}]"
        requestBody: "{\"name\":\"홍길동\"}"
        response: "{\"message\":\"성공\"}"
        description: 이 API는 예시 데이터를 처리합니다.
        httpMethod: GET
        apiSpecVersionId: 101
        version: 1
      properties:
        apiSpecVersionId:
          example: 101
          format: int64
          type: integer
        endpoint:
          example: "/api/v1/examples/{id}"
          type: string
        apiGroup:
          example: example
          type: string
        version:
          example: 1
          type: integer
        summary:
          example: 예시 스펙입니다.
          type: string
        description:
          example: 이 API는 예시 데이터를 처리합니다.
          type: string
        httpMethod:
          description: "HTTP 메서드 (GET, POST, PUT, PATCH, DELETE)"
          enum:
          - GET
          - POST
          - PUT
          - PATCH
          - DELETE
          type: string
        requestBody:
          example: "{\"name\":\"홍길동\"}"
          type: string
        queryParameters:
          example: "[{\"name\": \"id\", \"type\": \"integer\", \"required\": true}]"
          type: string
        pathParameters:
          example: "[{\"name\": \"exampleId\", \"type\": \"string\", \"required\"\
            : true}]"
          type: string
        response:
          example: "{\"message\":\"성공\"}"
          type: string
      type: object
    ApiSpecVersionCreatedResponse:
      example:
        apiSpecVersionId: 101
      properties:
        apiSpecVersionId:
          example: 101
          format: int64
          type: integer
      type: object
...
    PostSortEnumDto:
      description: "게시글 정렬 기준 (작성일, 조회수, 좋아요 수)"
      enum:
      - createdAt
      - viewCount
      - likeCount
      type: string
    PostOrderEnumDto:
      description: 정렬 방향 (오름차순 또는 내림차순)
      enum:
      - asc
      - desc
      type: string
    SearchTypeEnumDto:
      enum:
      - title
      - content
      - title_content
      type: string
    GetPostListResponse:
      example:
        pageable:
          isLastPage: true
          totalPages: 1
          listSize: 6
          isFirstPage: true
          totalElements: 5
        content:
        - createdAt: 2000-01-23T04:56:07.000+00:00
          dislikeCount: 1
          likeCount: 10
          postId: 101
          viewCount: 123
          title: 게시글 제목
          category: 공지
          commentCount: 5
          status: PENDING
        - createdAt: 2000-01-23T04:56:07.000+00:00
          dislikeCount: 1
          likeCount: 10
          postId: 101
          viewCount: 123
          title: 게시글 제목
          category: 공지
          commentCount: 5
          status: PENDING
      properties:
        content:
          items:
            $ref: '#/components/schemas/PostSummaryResponse'
          type: array
        pageable:
          $ref: '#/components/schemas/PageDto'
      type: object
...
    PostStatusEnumDto:
      description: "게시글 상태 (대기, 검토 중, 진행 중, 완료, 거부)"
      enum:
      - PENDING
      - REVIEWING
      - IN_PROGRESS
      - COMPLETED
      - REJECTED
      type: string
    CreatePostResponse:
      example:
        postId: 101
      properties:
        postId:
          example: 101
          format: int64
          type: integer
      type: object
    PostDetailResponse:
      example:
        author:
          role: USER
          profileImgUrl: https://example.com/profile.jpg
          nickname: john_doe
          userId: 550e8400-e29b-41d4-a716-446655440000
          username: Jaasdfas@fasdf.com
        dislikeCount: 1
        isUpdated: true
        likeCount: 10
        postId: 101
        title: 상세 게시글 제목
        content: 상세 게시글 내용입니다.
        commentCount: 5
        createdAt: 2000-01-23T04:56:07.000+00:00
        viewCount: 123
        category: 공지
        updatedAt: 2000-01-23T04:56:07.000+00:00
        status: PENDING
      properties:
        postId:
          example: 101
          format: int64
          type: integer
        title:
          example: 상세 게시글 제목
          type: string
        content:
          example: 상세 게시글 내용입니다.
          type: string
        category:
          example: 공지
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        viewCount:
          example: 123
          format: int64
          type: integer
        likeCount:
          example: 10
          format: int64
          type: integer
        dislikeCount:
          example: 1
          format: int64
          type: integer
        commentCount:
          example: 5
          type: integer
        isUpdated:
          example: true
          type: boolean
        status:
          $ref: '#/components/schemas/PostStatusEnumDto'
        author:
          $ref: '#/components/schemas/UserResponse'
      type: object
    VoteResponse:
      example:
        dislikeCount: 2
        likeCount: 12
      properties:
        likeCount:
          example: 12
          format: int64
          type: integer
        dislikeCount:
          example: 2
          format: int64
          type: integer
      type: object
...
    ApiSpecVersionUpdateRequest:
      example:
        summary: 예시 스펙입니다.
        pathParameters: "{ \"id\": \"123\" }"
        endpoint: "/api/v1/examples/{id}"
        queryParameters: "{ \"page\": \"1\", \"size\": \"10\" }"
        requestBody: "{ \"name\": \"홍길동\" }"
        response: JSON 응답 예시 문자열
        scrudProjectId: 101
        description: 이 API는 예시 데이터를 처리합니다.
        httpMethod: POST
      properties:
        scrudProjectId:
          example: 101
          format: int64
          type: integer
        endpoint:
          description: API 엔드포인트 경로
          example: "/api/v1/examples/{id}"
          type: string
        summary:
          description: API에 대한 간단한 설명
          example: 예시 스펙입니다.
          type: string
        description:
          description: API 상세 설명
          example: 이 API는 예시 데이터를 처리합니다.
          type: string
        response:
          description: 응답 예시(JSON 형태의 문자열)
          example: JSON 응답 예시 문자열
          type: string
        httpMethod:
          description: HTTP 메서드
          enum:
          - GET
          - POST
          - PUT
          - PATCH
          - DELETE
          example: POST
          type: string
        requestBody:
          description: 요청 바디 (JSON 형식 문자열)
          example: "{ \"name\": \"홍길동\" }"
          format: json
          type: string
        queryParameters:
          description: 쿼리 파라미터 (JSON 형식 문자열)
          example: "{ \"page\": \"1\", \"size\": \"10\" }"
          format: json
          type: string
        pathParameters:
          description: 경로 파라미터 (JSON 형식 문자열)
          example: "{ \"id\": \"123\" }"
          format: json
          type: string
      required:
      - endpoint
      - httpMethod
      type: object
...
    ApiSpecVersionStatusRequest:
      example:
        apiSpecStatus: AI_GENERATED
      properties:
        apiSpecStatus:
          description: "API 스펙 상태 (AI_GENERATED, AI_VISUALIZED, USER_COMPLETED)"
          enum:
          - AI_GENERATED
          - AI_VISUALIZED
          - USER_COMPLETED
          example: AI_GENERATED
          type: string
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: "JWT 인증 토큰을 사용합니다. 예: Bearer {token}"
      scheme: bearer
      type: http